22.04.28 12:20 환경변수 key, value 리스트로 관리 환경변수를 key(name)와 value로 구분해서 리스트로 관리하기로 함. 이에 따라 헤더파일에 들어가는 내용이 많아져 astree관련된 내용은 astree.h로 분리 close #16
