+++ sources/minishell.h
@@ -6,7 +6,7 @@
 /*   By: hyojlee <hyojlee@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/06 15:39:28 by hyojlee           #+#    #+#             */
-/*   Updated: 2022/04/27 21:26:30 by hyojlee          ###   ########.fr       */
+/*   Updated: 2022/04/28 12:20:15 by hyojlee          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,68 +22,23 @@
 # include <string.h>
 # include <sys/errno.h>
 # include "../libft/libft.h"
+# include "./astree.h"
 
 # define REPLACE 1
-
-typedef enum	e_std
-{
-	STDIN = 0,
-	STDOUT,
-	STDERR
-}	t_std;
-
-typedef enum	e_type
-{
-	TOKEN = -1,
-	BLANK = -2,
-	PIPE = '|',
-	LESS = '<',
-	GREAT = '>',
-	SQUOTE = '\'',
-	DQUOTE = '\"',
-	DLESS = -3,
-	DGREAT = -4,
-	REDIR = -5,
-	CMD = -6,
-	HEREDOC = -7
-}	t_type;
-
-typedef char	*t_data;
-
-typedef struct s_tok
-{
-	t_data	data;
-	t_type	type;
-	struct s_tok *next;
-}	t_tok;
-
-typedef struct s_tok_list
-{
-	t_tok	*head;
-	int		count;
-}	t_tok_list;
-
-typedef struct s_node
-{
-	t_data	data;
-	t_type	type;
-	struct s_node	*left;
-	struct s_node	*right;
-}	t_node;
-
-typedef struct s_astree
-{
-	t_node	*root;
-} t_astree;
 
 typedef struct s_info
 {
 	t_tok_list	*list;
 	t_astree	*tree;
-	char		**env;
 	t_list		*env_list;
 	int			exitcode;
 }	t_info;
+
+typedef struct s_enode
+{
+	char	*key;
+	char	*value;
+}	t_enode;
 
 /*
  * main.c
@@ -103,6 +58,7 @@ void	list_clear(t_tok_list *list);
 */
 void	tokenize(t_tok_list **list, char *str);
 int		ft_isblank(char c);
+
 /*
 ** quotecheck.c
 */
@@ -125,31 +81,25 @@ void	filename(t_info *info, int *idx);
 ** util.c
 */
 void	print_token(t_tok_list *list);
-void print_tree(t_node* root);
+void	print_tree(t_node* root);
 void	as_print(t_node *syntax);
 void	ft_clear(t_info *info);
 
 /*
-** astree.c
-*/
-
-void		insert_pipe_heredoc(t_astree *tree, t_node *node);
-t_node		*create_node(t_tok	*token);
-t_astree	*create_tree(void);
-void		insert_redir(t_astree *tree, t_node *node);
-void 		insert_path(t_astree *tree, t_node *node);
-void		insert_filename(t_astree *tree, t_node *node);
-void		tree_clear(t_astree *tree);
-
-/*
 ** semantic.c
 */
-int			chk_syntax(t_node *node);
+int		chk_syntax(t_node *node);
 
 /*
 ** replace_env.c
 */
-char	*get_env(char **env, char *name);
 void	replace_recur(t_info *info, t_node *node);
+
+/*
+** env_list.c
+*/
+void	env_preprocess(t_info *info, char **envp);
+char	*get_env(t_info *info, char *name);
+
 
 #endif
